+++ app/src/main/java/com/example/aquariumtracker/database/model/Measurement.kt
@@ -9,8 +9,8 @@ import androidx.room.PrimaryKey
         [ForeignKey(entity = Parameter::class,
         parentColumns = ["param_id"], childColumns = ["param_id"])])
 data class Measurement(
-    @PrimaryKey(autoGenerate = true) val measure_id: Int,
-    @ColumnInfo val param_id: Int,
+    @PrimaryKey(autoGenerate = true) val measure_id: Long,
+    @ColumnInfo val param_id: Long,
     @ColumnInfo val value: Double?,
     //@ColumnInfo val time: String,
     @ColumnInfo val time: Long
